+++ iowmongotools/__init__.py
@@ -3,15 +3,10 @@ __author__ = "Denis Ashcheulov"
 __version__ = "0.0.2"
 __status__ = "Planning"
 
-import sys
 import logging
 from iowmongotools import app
 
 logger = logging.getLogger(__name__)
-
-
-def hello():
-    return "Hello world!"
 
 
 class MongoCheck(app.App):
@@ -19,10 +14,9 @@ class MongoCheck(app.App):
     @property
     def default_config(self):
         config = super().default_config
-        config['logging'][0]['formatters']['console'].update(
-            {'format': '%(asctime)s %(levelname)s %(module)s: %(message)s'})
         config.update({
-            'efgbvwef': ('config.json', 'wefve')
+            'clusters': (('or', 'sc', 'eu', 'jp'), 'Set of a designation of clusters'),
+            'upload_dir': ('upload', 'Directory for uploads')
         })
         return config
 
@@ -31,4 +25,4 @@ class MongoCheckCli(MongoCheck, app.AppCli):
 
     def run(self):
         logger.debug('wefv')
-        sys.exit(0)
+        return 0